--- src/test/java/uk/org/tombolo/importer/phe/PhysicalActivityImporterTest.java
@@ -1,9 +1,0 @@
-package uk.org.tombolo.importer.phe;
-
-import uk.org.tombolo.AbstractTest;
-
-/**
- * Created by tbantis on 21/02/2018.
- */
-public class PhysicalActivityImporterTest extends AbstractTest {
-}
