+++ src/test/resources/datacache/TomboloData/uk.gov.phe/eb210218-c1b2-321f-932e-0b09568ca3b0.csv
@@ -1,0 +1,11 @@
+"Dataset","Age 16+"
+
+"Measure","Active","Active - 150-599 MVPA","Active - 600+ MVPA"
+"Geography","Adur","Allerdale","Amber Valley","Arun","Ashfield","Ashford","Aylesbury Vale","Babergh","Barking and Dagenham","Barnet","Barnsley","Barrow-in-Furness","Basildon","Basingstoke and Deane","Bassetlaw","Bath and North East Somerset","Bedford","Bexley","Birmingham","Blaby","Blackburn with Darwen ","Blackpool ","Bolsover","Bolton","Boston","Bournemouth ","Bracknell Forest ","Bradford","Braintree","Breckland","Brent","Brentwood","Brighton and Hove ","Bristol, City of, ","Broadland","Bromley","Bromsgrove","Broxbourne","Broxtowe","Burnley","Bury","Calderdale","Cambridge","Camden","Cannock Chase","Canterbury","Carlisle","Castle Point","Central Bedfordshire","Charnwood","Chelmsford","Cheltenham","Cherwell","Cheshire East","Cheshire West and Chester","Chesterfield","Chichester","Chiltern","Chorley","Christchurch","City of London","Colchester","Copeland","Corby","Cornwall","Cotswold","County Durham","Coventry","Craven","Crawley","Croydon","Dacorum","Darlington","Dartford","Daventry","Derby","Derbyshire Dales","Doncaster","Dover","Dudley","Ealing","East Cambridgeshire","East Devon","East Dorset","East Hampshire","East Hertfordshire","East Lindsey","East Northamptonshire","East Riding of Yorkshire","East Staffordshire","Eastbourne","Eastleigh","Eden","Elmbridge","Enfield","Epping Forest","Epsom and Ewell","Erewash","Exeter","Fareham","Fenland","Forest Heath","Forest of Dean","Fylde","Gateshead","Gedling","Gloucester","Gosport","Gravesham","Great Yarmouth","Greenwich","Guildford","Hackney","Halton","Hambleton","Hammersmith and Fulham","Harborough","Haringey","Harlow","Harrogate","Harrow","Hart","Hartlepool","Hastings","Havant","Havering","Herefordshire (County of)","Hertsmere","High Peak","Hillingdon","Hinckley and Bosworth","Horsham","Hounslow","Huntingdonshire","Hyndburn","Ipswich","Isle of Wight ","Isles of Scilly","Islington","Kensington and Chelsea","Kettering","King's Lynn and West Norfolk","Kingston upon Hull, City of","Kingston upon Thames","Kirklees","Knowsley","Lambeth","Lancaster","Leeds","Leicester","Lewes","Lewisham","Lichfield","Lincoln","Liverpool","Luton","Maidstone","Maldon","Malvern Hills","Manchester","Mansfield","Medway ","Melton","Mendip","Merton","Mid Devon","Mid Suffolk","Mid Sussex","Middlesbrough","Milton Keynes ","Mole Valley","New Forest","Newark and Sherwood","Newcastle upon Tyne","Newcastle-under-Lyme","Newham","North Devon","North Dorset","North East Derbyshire","North East Lincolnshire","North Hertfordshire","North Kesteven","North Lincolnshire","North Norfolk","North Somerset ","North Tyneside","North Warwickshire","North West Leicestershire","Northampton","Northumberland","Norwich","Nottingham City","Nuneaton and Bedworth","Oadby and Wigston","Oldham","Oxford","Pendle","Peterborough","Plymouth","Poole ","Portsmouth ","Preston","Purbeck","Reading ","Redbridge","Redcar & Cleveland","Redditch","Reigate and Banstead","Ribble Valley","Richmond upon Thames","Richmondshire","Rochdale","Rochford","Rossendale","Rother","Rotherham","Rugby","Runnymede","Rushcliffe","Rushmoor","Rutland","Ryedale","Salford","Sandwell","Scarborough","Sedgemoor","Sefton","Selby","Sevenoaks","Sheffield","Shepway","Shropshire","Slough ","Solihull","South Bucks","South Cambridgeshire","South Derbyshire","South Gloucestershire","South Hams","South Holland","South Kesteven","South Lakeland","South Norfolk","South Northamptonshire","South Oxfordshire","South Ribble","South Somerset","South Staffordshire","South Tyneside","Southampton ","Southend-on-Sea","Southwark","Spelthorne","St. Albans","St. Edmundsbury","St. Helens","Stafford","Staffordshire Moorlands","Stevenage","Stockport","Stockton-on-Tees","Stoke-on-Trent","Stratford-on-Avon","Stroud","Suffolk Coastal","Sunderland","Surrey Heath","Sutton","Swale","Swindon","Tameside","Tamworth","Tandridge","Taunton Deane","Teignbridge","Telford and Wrekin","Tendring","Test Valley","Tewkesbury","Thanet","Three Rivers","Thurrock","Tonbridge and Malling","Torbay ","Torridge","Tower Hamlets","Trafford","Tunbridge Wells","Uttlesford","Vale of White Horse","Wakefield","Walsall","Waltham Forest","Wandsworth","Warrington ","Warwick","Watford","Waveney","Waverley","Wealden","Wellingborough","Welwyn Hatfield","West Berkshire ","West Devon","West Dorset","West Lancashire","West Lindsey","West Oxfordshire","West Somerset","Westminster","Weymouth and Portland","Wigan","Wiltshire","Winchester","Windsor and Maidenhead ","Wirral","Woking","Wokingham ","Wolverhampton","Worcester","Worthing","Wychavon","Wycombe","Wyre","Wyre Forest","York "
+"Time period","2012 (Mid-January 2012 to Mid-January 2013)","2013 (Mid-January 2013 to Mid-January 2014)","2014 (Mid-January 2014 to Mid-January 2015)"
+
+"","","","Time period","Time period","Time period"
+"","","","2012 (Mid-January 2012 to Mid-January 2013)","2013 (Mid-January 2013 to Mid-January 2014)","2014 (Mid-January 2014 to Mid-January 2015)"
+"Active","Geography","Islington","61.3%","62.0%","66.2%"
+"Active - 150-599 MVPA","Geography","Islington","37.7%","35.6%","32.5%"
+"Active - 600+ MVPA","Geography","Islington","23.6%","26.4%","33.8%"
